Sync verifier
