+++ Cargo.toml
@@ -32,3 +32,4 @@ async-trait = "0.1"
 hex = "0.4"
 secrecy = "0.8.0"
 byteorder = "1.5.0"
+url = "2.5.4"
